fix: calibrate touch threshold for arduino-esp32 3.x semantics
In core 3.x touch readings rise on touch (2.x fell), so the legacy
default threshold of 40 sat far below the ~1400 untouched baseline,
making the interrupt condition permanently true: one phantom press at
boot and a dead button after. Auto-calibrate at construction: if the
configured threshold is at or below the untouched baseline, set it 20%
above the baseline. Polling comparison flipped to match (touched =
value above threshold) and the threshold widened to uint32_t.

+++ src/touch.h
@@ -39,7 +39,7 @@ class TouchButton {
 private:
   bool lastState;                       // Last stable state
   uint8_t pin;                          // Touch pin number
-  uint16_t threshold;                   // Touch threshold value
+  uint32_t threshold;                   // Touch threshold value
   volatile unsigned long lastPressTime; // Last state change time for debouncing
   volatile bool pressedFlag;            // Flag indicating button press detected
   unsigned long debounceTime;           // Configurable debounce time
@@ -54,7 +54,7 @@ public:
    * @param debounceMs The debounce time in milliseconds (default 100)
    * @param useInterrupt Whether to use interrupt mode (default false)
    */
-  TouchButton(uint8_t touchPin, uint16_t touchThreshold = 40, unsigned long debounceMs = 100, bool useInterrupt = false);
+  TouchButton(uint8_t touchPin, uint32_t touchThreshold = 40, unsigned long debounceMs = 100, bool useInterrupt = false);
 
   /**
    * @brief Handle touch button state
